chore: remove old files move new ones
