+++ src/classes/FountainDecoder.ts
