+++ src/classes/FountainEncoder.ts
