--- src/index.old.ts
@@ -1,46 +1,0 @@
-import { Decoder, IDecoder } from "./classes/Decoder.js";
-import { Encoder, IEncoder } from "./classes/Encoder.js";
-import { MultipartUr } from "./classes/MultipartUr.js";
-import {
-  createUrTranscoder,
-  createMultipartUrTranscoder,
-  createFountainUrTranscoder,
-} from "./classes/ngraveTranscoder.js";
-import { Ur } from "./classes/Ur.js";
-import { UrDecoder } from "./classes/UrDecoder.js";
-import { UrEncoder } from "./classes/UrEncoder.js";
-import { UrMultipartEncoder } from "./classes/UrMultipartEncoder.js";
-import { UrMultipartDecoder } from "./classes/UrMultipartDecoder.js";
-import UrFountainDecoder from "./classes/UrFountainDecoder.js";
-import UrFountainEncoder from "./classes/UrFountainEncoder.js";
-import { RegistryItem, registryItemFactory, RegistryItemClass } from "./classes/RegistryItem.js";
-import { UrRegistry } from "./registry.js";
-import { CborEncoding } from "./encodingMethods/CborEncoding.js";
-
-
-export {
-  Ur,
-}
-
-// export {
-//   UrRegistry,
-//   RegistryItem,
-//   registryItemFactory,
-//   RegistryItemClass,
-//   Ur,
-//   MultipartUr,
-//   createUrTranscoder,
-//   createMultipartUrTranscoder,
-//   createFountainUrTranscoder,
-//   IDecoder,
-//   Decoder,
-//   UrDecoder,
-//   CborEncoding,
-//   UrMultipartDecoder,
-//   UrFountainDecoder,
-//   IEncoder,
-//   Encoder,
-//   UrEncoder,
-//   UrMultipartEncoder,
-//   UrFountainEncoder,
-// };
