+++ src/index.ts
@@ -1,22 +1,24 @@
-import { Ur } from "./new_classes/Ur.js"
+import { Ur } from "./classes/Ur.js"
 import { UrRegistry } from "./registry.js"
 import { registryItemFactory } from "./classes/RegistryItem.js"
-import { UrFountainEncoder } from "./new_classes/UrFountainEncoder.js"
-import { UrFountainDecoder } from "./new_classes/UrFountainDecoder.js"
-import { FountainEncoder } from "./new_classes/FountainEncoder.js"
-import { FountainDecoder } from "./new_classes/FountainDecoder.js"
-
-// Import types
-import type { RegistryItem, RegistryItemClass } from "./classes/RegistryItem.js"
+import { UrFountainEncoder } from "./classes/UrFountainEncoder.js"
+import { UrFountainDecoder } from "./classes/UrFountainDecoder.js"
+import { FountainEncoder } from "./classes/FountainEncoder.js"
+import { FountainDecoder } from "./classes/FountainDecoder.js"
+import { defaultEncoders, dataPipeline } from "./encodingMethods/index.js"
 
 export {
   UrRegistry,
-  RegistryItem,
-  RegistryItemClass,
   registryItemFactory,
   Ur,
   UrFountainDecoder,
   UrFountainEncoder,
   FountainDecoder,
-  FountainEncoder
+  FountainEncoder,
+  defaultEncoders,
+  dataPipeline,
 }
+
+// Export types
+import type { RegistryItem, RegistryItemClass } from "./classes/RegistryItem.js"
+export type { RegistryItem, RegistryItemClass }
