+++ tests/fountain.test.ts
@@ -1,10 +1,10 @@
 import { makeMessage } from "../src/helpers/utils";
 import { uint8ArrayToHex } from "uint8array-extras";
-import { Ur } from "../src/new_classes/Ur";
-import { FountainEncoder } from "../src/new_classes/FountainEncoder"
-import { FountainDecoder } from "../src/new_classes/FountainDecoder"
-import { UrFountainEncoder } from "../src/new_classes/UrFountainEncoder";
-import { UrFountainDecoder } from "../src/new_classes/UrFountainDecoder";
+import { Ur } from "../src/classes/Ur";
+import { FountainEncoder } from "../src/classes/FountainEncoder"
+import { FountainDecoder } from "../src/classes/FountainDecoder"
+import { UrFountainEncoder } from "../src/classes/UrFountainEncoder";
+import { UrFountainDecoder } from "../src/classes/UrFountainDecoder";
 
 
 function make_message_ur(len: number, seed?: string): Ur {
